@@ -10,19 +10,21 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "12.2.8",
-    "@angular/common": "12.2.8",
-    "@angular/compiler": "12.2.8",
-    "@angular/core": "12.2.8",
-    "@angular/forms": "12.2.8",
-    "@angular/platform-browser": "12.2.8",
-    "@angular/platform-browser-dynamic": "12.2.8",
-    "@angular/router": "12.2.8",
+    "@angular/animations": "13.0.0-next.15",
+    "@angular/common": "13.0.0-next.15",
+    "@angular/compiler": "13.0.0-next.15",
+    "@angular/core": "13.0.0-next.15",
+    "@angular/forms": "13.0.0-next.15",
+    "@angular/platform-browser": "13.0.0-next.15",
+    "@angular/platform-browser-dynamic": "13.0.0-next.15",
+    "@angular/router": "13.0.0-next.15",
     "@auth0/angular-jwt": "~5.0.2",
     "@microsoft/applicationinsights-web": "~2.7.0",
     "@ngrx/effects": "~12.4.0",
     "@ngrx/store": "~12.4.0",
     "@ngrx/store-devtools": "~12.4.0",
+    "@ngx-translate/core": "^13.0.0",
+    "@ngx-translate/http-loader": "^6.0.0",
     "@popperjs/core": "^2.10.2",
     "angularx-social-login": "~4.0.1",
     "bootstrap": "^5.1.1",
@@ -31,9 +33,9 @@
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "12.2.8",
-    "@angular/cli": "12.2.8",
-    "@angular/compiler-cli": "12.2.8",
+    "@angular-devkit/build-angular": "13.0.0-next.9",
+    "@angular/cli": "13.0.0-next.9",
+    "@angular/compiler-cli": "13.0.0-next.15",
     "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
     "jasmine-auto-spies": "~6.9.5",
@@ -44,6 +46,6 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "typescript": "~4.3.5"
+    "typescript": "~4.4.2"
   }
 }
