@@ -20,46 +20,11 @@ export class AuthGuard implements CanActivate {
             return this.authService.hasCurentUserAMatchingRole( route.data.roles);
         }
 
-        const isRefreshSuccess = await this.tryRefreshingTokens(token);
+        const isRefreshSuccess = await this.authService.tryRefreshingTokens(token);
         if (!isRefreshSuccess) {
             this.router.navigate(['login']);
         }
 
         return this.authService.hasCurentUserAMatchingRole( route.data.roles);
-    }
-
-    private async tryRefreshingTokens(token: string): Promise<boolean> {
-        // Try refreshing tokens using refresh token
-        const refreshToken: string = localStorage.getItem('refreshtoken');
-
-        if (!token || !refreshToken) {
-            return false;
-        }
-
-        const credentials = JSON.stringify({
-            accessToken: token,
-            refreshToken: refreshToken,
-        });
-
-        let isRefreshSuccess: boolean;
-        try {
-            const response = await this.http
-                .post('/token/refresh', credentials, {
-                    headers: new HttpHeaders({
-                        'Content-Type': 'application/json',
-                    }),
-                    observe: 'response',
-                })
-                .toPromise();
-            // If token refresh is successful, set new tokens in local storage.
-            const newToken = (<any>response).body.accessToken;
-            const newRefreshToken = (<any>response).body.refreshToken;
-            localStorage.setItem('jwt', newToken);
-            localStorage.setItem('refreshToken', newRefreshToken);
-            isRefreshSuccess = true;
-        } catch (ex) {
-            isRefreshSuccess = false;
-        }
-        return isRefreshSuccess;
     }
 }
